+++ src/Modules/Workflows/Module.php
@@ -134,6 +134,8 @@ final class Module implements InitializableInterface
         $this->options = $options;
         $this->email = $email;
 
+        $this->registerHooks();
+
         /*
          * We initialize the engine in the constructor because it requires
          * the init hook has not been fired yet. The initialize method runs in the init hook.
@@ -148,6 +150,7 @@ final class Module implements InitializableInterface
     public function initialize()
     {
         $this->initializeControllers();
+
     }
 
     private function initializeControllers()
@@ -211,4 +214,25 @@ final class Module implements InitializableInterface
 
         $this->workflowEngine->start();
     }
+
+    private function registerHooks()
+    {
+        $this->hooks->addFilter(
+            HooksAbstract::FILTER_WORKFLOW_ROUTE_VARIABLE,
+            [$this, 'filterWorkflowRouteVariable'],
+            10,
+            2
+        );
+    }
+
+    public function filterWorkflowRouteVariable($variableName, $dataSource)
+    {
+        // The execution_id variable is deprecated.
+        if ($variableName === 'global.execution_id') {
+            return 'global.run_id';
+        }
+
+        return $variableName;
+    }
+
 }